Move funcs to grammar
